Extract HomeChecker from HomeView
